pyproject.toml
[project]
name = "features-analyzer"
version = "0.1.0"
description = "Machine learning features analyzer dashboard"
keywords = ["machine learning", "features", "ia", "proto", "dashboard"]
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
  { name = "Zaú Júlio A. Galvão, Main developer", email = "zaujulio.dev@gmail.com" },
]
maintainers = [
  { name = "Zaú Júlio A. Galvão, Main developer", email = "zaujulio.dev@gmail.com" },
]

# For a list of valid classifiers, see https://pypi.org/classifiers/
classifiers = [
  "Development Status :: 1 - Planning",
  "Environment :: X11 Applications :: GTK",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Topic :: Software Development :: Build Tools",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.12",
]

[project.urls]
"Homepage" = "https://github.com/ZauJulio/FeaturesAnalyzerGTK"
"Bug Reports" = "https://github.com/ZauJulio/FeaturesAnalyzerGTK/issues"
"Funding" = "https://github.com/sponsors/ZauJulio"
"Say Thanks!" = "https://saythanks.io/to/ZauJulio"
"Source" = "https://github.com/ZauJulio/FeaturesAnalyzerGTK/"

[tool.poetry]
name="features-analyzer"
package-mode = false

[tool.poetry.dependencies]
matplotlib = "==3.10.0"
numpy = "==2.2.1"
pandas = "==2.2.3"
pygobject = "==3.50.0"
python = "^3.12"
seaborn = "^0.13.2"
scikit-learn = "^1.6.0"
python-dotenv = "^1.0.1"
pydantic = "^2.10.4"
tinydb = "^4.8.2"
scipy = "^1.14.1"
python-i18n = "^0.3.9"
joblib = "^1.4.2"

[tool.poetry.group.dev.dependencies]
jinja2 = "^3.1.5"
lxml = "*"
poetry-plugin-export = "^1.0.6"
pre-commit = "*"
pygobject-stubs = "==2.12.0"
ruff = "*"
pandas-stubs = "^2.2.3.241126"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
